+++ lib/tracker_coin_package.dart
@@ -7,7 +7,7 @@ import 'package:http/http.dart' as http;
 class TrackerCoinPackage {
   /// This is the main method of the package [fetchCoin] where we use the http package to call an api.
   /// here we access coingecko free api to check coin information.
-  Future<List<Coin>> fetchCoin() async {
+  Future<List<CoinModel>> fetchCoin() async {
     coinList = [];
     final response = await http.get(Uri.parse(
         'https://api.coingecko.com/api/v3/coins/markets?vs_currency=usd&order=market_cap_desc&per_page=100&page=1&sparkline=false'));
@@ -15,15 +15,19 @@ class TrackerCoinPackage {
     ///Basic http call.
     if (response.statusCode == 200) {
       List<dynamic> values = [];
+      // We decode the value.
       values = json.decode(response.body);
+      //We check the list.
       if (values.isNotEmpty) {
         for (int i = 0; i < values.length; i++) {
+          //We add the call.
           if (values[i] != null) {
             Map<String, dynamic> map = values[i];
-            coinList.add(Coin.fromJson(map));
+            coinList.add(CoinModel.fromJson(map));
           }
         }
       }
+      //We return our coin list.
       return coinList;
     } else {
       throw Exception('Failed to load coins');
@@ -31,10 +35,10 @@ class TrackerCoinPackage {
   }
 }
 
-/// [Coin] is the model that we are handling inside our package.
+/// [CoinModel] is the model that we are handling inside our package.
 /// We will need this in order to have our [fetchCoin] method working.
-class Coin {
-  Coin({
+class CoinModel {
+  CoinModel({
     required this.name,
     required this.symbol,
     required this.imageUrl,
@@ -42,7 +46,7 @@ class Coin {
     required this.change,
     required this.changePercentage,
   });
-
+//Atributes we want to have in our Model
   String name;
   String symbol;
   String imageUrl;
@@ -50,8 +54,8 @@ class Coin {
   num change;
   num changePercentage;
 
-  factory Coin.fromJson(Map<String, dynamic> json) {
-    return Coin(
+  factory CoinModel.fromJson(Map<String, dynamic> json) {
+    return CoinModel(
       name: json['name'],
       symbol: json['symbol'],
       imageUrl: json['image'],
@@ -64,4 +68,4 @@ class Coin {
 
 /// This is the list that we will be displaying when using the package.
 
-List<Coin> coinList = [];
+List<CoinModel> coinList = [];